refactor: seedWork repository was refactored
