+++ src/@seedwork/errors/in-memory/not-found.error.ts
