+++ src/Category/domain/repository/category.repository.ts
@@ -1,0 +1,6 @@
+import { RepositoryInterface } from "@seedwork/domain/repository/repository-contracts";
+import { Category } from "../entity/category";
+
+export interface CategoryRepository extends RepositoryInterface <Category> {
+
+}
