+++ src/Category/infra/repository/category-in-memory.repositorie.ts
@@ -1,0 +1,8 @@
+import { Category } from "../../domain/entity/category";
+import InMemoryRepository, {} from "../../../@seedwork/domain/repository/repository";
+import { CategoryRepository } from "../../domain/repository/category.repository";
+
+class CategoryInMemoryRepository
+  extends InMemoryRepository<Category>
+  implements CategoryRepository {
+}
